[camera] Fix sherlock ISP IRQ mode setting
This change updates the sherlock board configuration to correct the ISP
IRQ mode to LEVEL_HIGH. This reflects its default (and current) setting
in the ISP. Prior to this change, if the ISP fell behind on processing,
the presence of buffered frame data would hold the GIC bit high despite
being reset by the ISR, preventing the interrupt from signaling. The
only reason this worked at all before is because when the ISP is gated
by the sensor speed, level-mode interrupts also act as pulses.

Bug: b/199321382
Test: Verified sherlock camera no longer stalls after heavy system load
Change-Id: Ie9e80a721a538fe7776d7537c98a7f9ab66c0398
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/642643
Reviewed-by: Marcus Jang <marcusjang@google.com>
Reviewed-by: Gurjant Kalsi <gkalsi@google.com>
Commit-Queue: Matt Sandy <msandy@google.com>