[ftl] Don't timeout after 60 seconds
Timing out after 60 seconds is problematic at the moment because it's
not possible to be certain that the callback will never be called or
that the operation has been cancelled.  The first will almost certainly
result in memory corruption and the latter has the potential to cause
confusion to higher layers.

Bug: 87653
Change-Id: I117c54b22a360b05cfa1cca7f5d70f7cf8954f31
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/643903
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Commit-Queue: Chris Suter <csuter@google.com>