[kernel][object] Clean up MSI object names
Some transitions since implementation of the original MSI design and
RFCs led to naming semantics that needed to be cleaned up, and object
usage that was unclear when libzx code was generated. For instance, FIDL
tools typically wanted to generate references to zx::msi_allocation,
since technically according to our sources that was correct despite
libzx settling on zx::msi. Ultimately, this was my fault and this is an
attempt to fix it.

MsiDispatcher -> MsiInterruptDispatcher
- Despite the slightly redundant name, the new name makes it clear that
  this is a derived InterruptDispatcher for use with MSIs.

MsiAllocationDispatcher -> MsiDispatcher
- From the perspective of userspace we simply have handles to msi
  blocks, and zx::msi. Having zx::msi and MsiAllocation Dispatcher was
  causing issues between generated FIDL bindings and handle types, so
  this simplifies it such that MsiDispatcher and MsiInterruptDispatchers
  represent everything necessary, and from the perspective of handle
  typing and libzx we simply treat MsiInterruptDispatchers as
  interrupts. For a similar reason, the MSII object type has been
  removed since it was never actually used and contributed to technical
  debt.

Change-Id: I75651741c0ff4f3a97eec29893f2cf8889701211
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/645101
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
API-Review: Adam Barth <abarth@google.com>
Commit-Queue: Christopher Anderson <cja@google.com>