[fdf][compat] Store Driver in Device
The zx_driver_t* pointer that a DFv1 driver hands to us
is global across the drivers of that type in a driver host.

We cannot use it to add a device, since it will add the
device to the most recently created driver of that type,
instead of adding it to the correct driver.

This is tested by having two instances of a driver each
add a device with the same name in the same driver host.

This also adds a test to make sure globals work correctly
within a DFv1 driver.

Bug: 94230

Change-Id: Ia2936ea0c96283049ac61ef6e7462a84cc43dea4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/648041
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: David Gilhooley <dgilhooley@google.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>