[fdf] Fix UB while removing node.
Node::Remove() would call Remove() on all of its children, which would
mutate the children list of the parent. This would result in undefined
behaviour because the vector was being iterated on when it was mutated.

This CL switches the children_ member to be a std::list, and slightly
modifies the Remove() loop so that we don't trigger UB.

Bug: 94561
Change-Id: I4b476cb1f1598bde8d1de37f97b3196760a19dfc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/651286
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Simon Shields <simonshields@google.com>