[fdf] Use eager drivers in driver index
The driver index will mark all fallback drivers as non-fallback drivers
if they are marked as eager. The driver index will read the boot
argument `devmgr.bind-eager` in order to determine which drivers are
eager. `devmgr.bind-eager` is a string that is a comma-separated list of
driver component URL's.

Bug: 92329
Change-Id: I816f627be7c0f1fc9e9f8b8b342698ed296c3429
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/646703
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Nick Eaton <nickeaton@google.com>