[devices] Use non-discardable fidl::Result
fidl::Result will soon be marked [[nodiscard]], making it necessary to
use the value. For the sake of this CL, the result is cast to void,
keeping the existing behavior.

Note: the first version of this CL checked that the result's status
was ZX_OK, but this failed on CQ. It may be useful to follow up on
this CL and diagnose the reason for it.

Bug: 94739
Change-Id: I9e4d431b707fe09e73016ff105bee7a5ae349222
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/653016
Fuchsia-Auto-Submit: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Sam Hansen <hansens@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>