[devices] Make ExposedInspector a class
This CL changes the `driver::ExposeInspect` function
to instead be a class. This was done because after changing
its usage from `service::OutgoingDirectory` to
`component::OutgoingDirectory`, another object (a vfs) associated
with the diagnostics directory needed to be kept alive.

For the convenience of clients of this function, it was better served
to wrap the internal objects in a class instead of exporting it
as a tuple<zx::Vmo, fs::ManagedVfs>.

Change-Id: Ideee38df5ec26bc3c25fd89624cde3d2fda1b0f8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/658504
Reviewed-by: Adam Perry <adamperry@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Yaneury Fermin <yaneury@google.com>