[usb][xhci] fix device state locking and check for disconnect
Originally we were getting the transaction_lock for a certain device
state to check if it is currently disconnecting, releasing the lock then
re-acquiring the lock to do something else. In between releasing the
lock and re-acquiring the lock, the disconnecting state may have
changed. Here we modify the code so that the lock is held until the end
of the operation so that we can be sure that the the device is not
disconnecting.

This CL also adds checks to disconnecting for device states in some
places that were originally missing.

Test: tested on Atlas

Change-Id: I61f98d553e6bc1ea666358c821141c7652a92268
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/656575
Reviewed-by: Sam Hansen <hansens@google.com>
Commit-Queue: Ruby Zhuang <rdzhuang@google.com>