[usb][hub] fix unbind deadlock
Originally we were deadlocking because the completion event was never
being signaled when returning from RequestQueue in ControlOut on a
failure, resulting in a deadlock.

Here we also fix the problem of the thread trying to join even when the
thread is not joinable.

Bug:94220

Test:Tested on Atlas

Change-Id: I9ed5b87ee881ee27a3039817edbc84758e967b22
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/657447
Reviewed-by: Sam Hansen <hansens@google.com>
Commit-Queue: Ruby Zhuang <rdzhuang@google.com>