[fdf][compat] Lazily connect to ProfileProvider.
Eagerly connecting causes a lot of log spam if a driver component
doesn't have the ProfileProvider routed to it, and is confusing.
Drivers generally don't make more than 1 call to the profile provider
anyway, so just connect to the profile provider when a profile is
requested.

Bug: 91608
Change-Id: I582e5cc6c15064419eed16ca14c83f35173a8e28
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/660548
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Simon Shields <simonshields@google.com>