[fdf] Driver manager supports device groups
Implement support for device groups in driver
manager. The Coordinator adds the device group by
first adding it to the driver index through the
driver loader, before passing the device group to
the BindDriverManager for binding the nodes.

The BindDriverManager creates a CompositeDevice
for the device group and maps it according to the
topological path.

Bug: 91510
Tests: Add unit test to driver-manager-tests
Change-Id: Ia78ad96f424d8df74388c0ec7851e21d64e9ac94
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/659263
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Sarah Chan <spqchan@google.com>