[ddk][fdf] Remove device_get_dispatcher
This was always a temporary measure. Now that we have the driver runtime
available in DFv1 driver hosts, we can use the dispatcher returned by it
instead of manage async loops in the driver host.

As a consequence we no longer spawn a thread per device. These were
largely unused anyways and were a waste of resources.

One outstanding change that is still necessary is a mechanism by which
we can cancel all outstanding tasks for a given device. This behavior is
equivlanet to what we had before, so it's no worse than before.

Fixed: 95885
MULTIPLY: driver_host_tests
Change-Id: I060f0ef7a13a6d7d6772efb6f905f137d4628d2f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/660424
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>