[driver_runtime] Fix DispatcherCoordinator race condition.
When calling the driver notify shutdown handler (usually will be
the driver host), we should not remove the driver_state object
until the handler has completed.

This is as we do not want the shutdown handler to be running,
while another thread thinks the all_drivers_destroyed condition
is true. This could happen if someone calls
fdf_internal_wait_until_all_dispatchers_destroyed or
the last dispatcher of the last driver is destroyed.

Bug: 87162

Change-Id: Ia87adbfa41d21b7cf0189042cbde761a9d49d2a7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669028
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>