[driver_runtime] Track threads
We centrally track how many threads we have spawned, avoiding spawning
new ones unnecessarily. We ideally want to reset back down to 1 thread
when possible, but that is tricky as it requires joining all threads
when we are running from within one of those threads being joined.

A Reset method was added which can be called from non-driver runtime
managed threads to reset the number of threads back to 1.

Additionally we clamp the overall number of threads we may spawn to 10
in order to reduce resource usage. This is a temporary solution which
needs to be solved with something more robust in the future.

MULTIPLY: driver_runtime_tests
Tested: fx test driver_runtime_tests -- --gtest_repeat=1000 \
           --gtest_break_on_failure
Change-Id: Iea8fb4c3f0c89e9f54455a80e16e61d1a67571da
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/651170
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>