[fdf] Mark if a matched driver is a fallback
Driver index will include a new boolean field in `MatchedDriverInfo`
called `is_fallback` that will denote whether or not the driver is a
fallback driver. This will be used in a future commit by the driver
loader in order to determine if a matched driver is a fallback driver.

Bug: 92329
Change-Id: I11d49fe624737da4f50acc9ad52b7319fdbe9d24
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669043
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
API-Review: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Nick Eaton <nickeaton@google.com>