[usb][ums] Disable ums-integration-test
The integration test is flaky, and we're having a hard time resolving
the race. Disabling the test to unwedge the tree while we sort out
what's going on with the virtual bus. UMS doesn't see a lot of delta, so
I'm ok turning off this test for the time being.

Bug: 98566
Change-Id: Ib316d3b79e1fa68505d291f2bea8527bc9a89298
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/671236
Reviewed-by: Ruby Zhuang <rdzhuang@google.com>
Commit-Queue: Sam Hansen <hansens@google.com>