[fdf][bind] Add instruction enums to DecodeBindRules
This CL updates DecodedBindRules so that it stores
the decoded instructions. It modifies
InstructionVerification so that in addition
to verifying the instruction bytecode, it also
converts it into a set of enums representing the
instruction.

This will make it easier to process the instructions
for matching and debugging.

Bug: 93365
Tests: Updated existing tests with the decoded
       instructions
Change-Id: I1f2a4ab9965855703c205538c702e6acc392662a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/671530
Reviewed-by: Novin Changizi <novinc@google.com>
Commit-Queue: Sarah Chan <spqchan@google.com>