Reland "[fdf] Set fallback and eager drivers with new way"
This is a reland of commit c40c967f13e978b00c65989379d7052eb818299d.
Make the driver loader use the `is_fallback` field returned by
MatchDriversV1 when considering if a driver is a fallback driver when
matching drivers to a device.

Originally, the driver loader checked if a driver's version began with
'*' in order to determine if a driver is a fallback driver. Now, the
driver index will check if a driver is a fallback driver instead and
will do so by checking if a driver's component manifest's `fallback`
field is set to true. The driver loader cannot check this. Making
fallback drivers use the new method without first updating the driver
loader to rely on the driver index to identify fallback drivers resulted
in unexpected behavior and is the reason why the original CL was
reverted.

Original change's description:
> [fdf] Set fallback and eager drivers with new way
>
> Set the `fallback` field in all fallback drivers' CML files to true.
> Remove '*' from fallback drivers' version.  Identify eager drivers using
> their component URL and not their name.  This is to avoid incorrect
> ordering of drivers returned by the driver index and processed by the
> driver loader.
>
> Bug: 92329
> Change-Id: I5842df7eb42620af21f89ab80a8f1d90f7d4a32f
> Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/654862
> Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
> Reviewed-by: Yilong Li <liyl@google.com>
> Reviewed-by: David Gilhooley <dgilhooley@google.com>
> Reviewed-by: Simon Shields <simonshields@google.com>
> Commit-Queue: Nick Eaton <nickeaton@google.com>

Bug: 92329
Change-Id: If1dc9b5397c6176f6618ac82b55b8ce82e6800ab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/667215
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Nick Eaton <nickeaton@google.com>