[driver_runtime] Improve safety of dispatcher passed to driver.
Instead of destroying the initial dispatcher passed to the driver
in the shutdown callback, we will destroy it when the driver
object is destructed.

This is as we will be passing an unowned dispatcher to the driver,
and it is better to guarantee the dispatcher will be valid for the
lifetime of the driver, otherwise the driver could attempt to
access an invalid pointer after the dispatcher is shutdown, such
as in its Stop hook.

Bug: 87162

Change-Id: Ia3b255167c0cb9f2ea3bb33e4be36dccb2a232d2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/671472
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>