[fdf] Support directory composite routing
Add support for composite devices for our directory routing.
Service instance aggregation is not supported yet, so we cannot
use services. But we can route subdirectories from parents to
children.

A directory capability name with a '-' represents a service
instance with the pattern:

 SERVICE_NAME-INSTANCE_NAME

If we are routing to a composite device, and we see a "default"
service instance being routed from a parent, that instance will
be renamed to the composite node name.

The primary composite node will be routed a second time as the
'default' instance.

TEST: fx test composite_test
Bug: 93678

Change-Id: I3052cc0ba2926a28ffecf8d594c4c5bcb0b33fcd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/673982
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: David Gilhooley <dgilhooley@google.com>