[pci] Don't assert if root_ isn't allocated in dtor
OOM testing demonstrated a situation where the Bus ctor path fails to
allocate the root node. In this case we properly assert that it wasn't
allocated in the dtor. This change switches over to verifying
allocations via alloc checker instead, and no longer asserts in the case
that root_ doesn't exist when we tear down the bus.

Bug: 99076
Change-Id: I1e1b5448c812047530e097231e6b22272da33119
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/674357
Reviewed-by: Andres Oportus <andresoportus@google.com>
Commit-Queue: Christopher Anderson <cja@google.com>